Разбил UI по файликам, Добавил Tab и разделение на хорошие и плохие привычки, починил разметку фильтров
